[fpga] Add transform to include explicit actor after each delay to initialize fifo and be included in the fifo analysis
