+++ plugins/org.preesm.algorithm/resources/templates/xilinxCodegen/template_top_kernel_fpga.cpp
@@ -52,10 +52,6 @@ $PREESM_INCLUDES
 
 $PREESM_SPECIAL_ACTORS
 
-// Wrappers to init delays
-
-$PREESM_DELAY_WRAPPERS
-
 // Wrapper to init actors
 
 $PREESM_INIT_WRAPPER
